Separate test_filter_metadata and test_filter_stats
